sub-second display timing: spline subdivision + RAF cursor
- Float32Array smoothed sidecar carries fractional-second t (deck.gl GPU
  attribute is float32 anyway; storing baseline-relative seconds gives
  ~0.06s precision over a 10-day window)
- Lower SPLINE_SUB_MIN_DT_S=0.5 + SPLINE_SUB_TARGET_DT_S=0.5 so 1Hz input
  data gets Hermite-subdivided to 2Hz (currently — easy to retune)
- useInterpolatedNow hook drives a 10Hz lag-based cursor: display sits 10s
  behind latestUpdate, advances at wall-clock rate so incoming WebSocket
  updates land ahead of the cursor and the advance is continuous
- selectAllPositions interpolates the marker position from the smoothed
  sidecar so the IconLayer dot slides along the same curve as the trail
- ogntripslayer picking reads from getTimestamps.value and adds
  referenceDate back to keep tooltip in epoch-seconds

+++ lib/types.ts
@@ -327,7 +327,11 @@ export interface SmoothedDeck {
     positions: Float32Array;
     indices?: Uint32Array;
     agl: Int16Array;
-    t: Uint32Array;
+    // Display-side time: fractional seconds since referenceDate. Float32 has
+    // ~0.06s precision at the 10-day baseline-relative scale, plenty for the
+    // TripsLayer animation cursor. Picking adds referenceDate back to recover
+    // epoch-seconds (lib/react/ogntripslayer.ts).
+    t: Float32Array;
     climbRate: Int8Array;
     // For each smoothed vertex, the index of the anchor it is emitted FOR
     // (the bracket's END anchor — inner vertices share the end anchor's
@@ -374,7 +378,9 @@ export type SortKey =
 
 export interface DisplayPilotTrackData extends PilotTrackData {
     deckAdditional: {
-        tr: Uint32Array;
+        // Float32 fractional seconds-from-referenceDate. Built in
+        // lib/react/deckvh.ts; passed straight to TripsLayer's getTimestamps.
+        tr: Float32Array;
         climb: Uint8Array;
         aheight: Uint8Array;
     };
